Better fix for the 'uao_ddl/uao_allalter' failure.
Commit f8d3dac1c18619f3365c9617edf108a1a4632235 changed the expected
output to accept the new result you get with multiple set-returning
functions in target list, but that made the comment above one of the
queries false: the query with FULL JOIN no longer returned the same number
of rows as the other queries. So instead of accepting the difference,
change the INSERT command so that it inserts the same set of rows as
before, but without using SRFs in the target list.