Update uao_ddl tests to use the intended access method
There is still an ongoing discussion regarding the environment set and the
storage options in relation to the ao/aoco related syntax. However, with the
current syntax these tests where not doing what they ment to do. As a result in
several cases, a row orientated relation would be created where a column
oriented was expected.

Update the tests to actually test what they set up to test. If the ongoing
discussion changes the behaviour, then let us update the tests again. On the
meantime we eliminate false positive tests.