Handle difference in namespace for EXCHANGE PARTITION
It's starting to get inefficient, as need to generate many stmts for
EXCHANGE PARTITION and involves dispatch for each of them. Since
didn't find way to rename table and alter schema for the same in
single stmt the number has gone up from 5 dispatch per exchange to 8
dispatches if namespace differ between tables.

It's functional but not efficient so far. In long run we can evaluate
how to make it efficient. I still wish to keep the legacy syntax
mapping only on QD and QEs just executing upstream way. So, prefer to
avoid dispatching EXCHANGE directly to segments.