Reinstate usage of RESERVED_SEGNO for appendoptimized relations
RESERVED_SEGNO is special because it is inserted as a regular tuple
(all other segnos are inserted as frozen tuples) into gp_fastsequence
relation.  This allows leveraging MVCC for clean up.  See Greenplum
commit a34214e19c24 for more details.

In current master, the caller of the insert API knows when to use
RESERVED_SEGNO.  The new table AM API forces us to hide this AM
specific detail from the callers.  Therefore, we have to infer the use
case at a much later point, where we don't have any information about
the callers.  E.g. CTAS, alter table rewrite or create and insert in
the same transaction.

This patch removes special status for segno = 1 because there is no
apparent valid use case where it needs the special status.  In master,
segno = 1 is used for same transaction create and insert case.  That
is no longer the case.  RESERVED_SEGNO is also used for same
transaction create and insert.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>