Fix bogus ERRCODE_TRIGGERED_DATA_CHANGE_VIOLATION during delete on AO relations
Delete needs to differenciate between a tuple that has been deleted by the
current command, or by a later command in the current transaction. The second
case indicates a BEFORE trigger. This check is performed outside of the tableam
api.

AO relations do not support BEFORE triggers and this check becomes superfluous.
In the case of a tuple being deleted in the current command, set up the state as
such to satisfy the callers expectations.

This hack is preferrable than leaking the relation's access method in the
caller.