Fix Error message for CTAS AO
We can not add RELOPT_KIND_HEAP in AO relopt, it will give an
inappropriate error message if CTAS a AO illegal. However, if we rid of
this tag, AOCO table will meet some problem when it create.

Rid of this tag first, will give a new way to recognize relopt when
AOCO merge start.