Reinstate table am handler separation
Although appendoptimized and aoco access methods will most probably share a lot
of code, they remain distinct access methods. Their handler should reflect that.
Make the handler functions static.

In the case that functionality is to be shared, then a call to a shared function
can be made. We can assess the situation once the access are more functional and
discuss the directory location for these functions.

Co-authored-by: Asim R P <apraveen@pivotal.io>