Align psql front end with new catalog entries discribing table access method
In greenplum we depended on relstorage column to distinguish between heap and
appendoptimized (row and column) relations. With the introduction of the table
am api in v12 of postgres, the greenplum specific relstorage column was removed.

Teach \d and \d+ to use the new or old method depending on the server version
and print the relevant info.

In the new method, it is possible that the desribe output is different,
especially in tests, due to the newly added HIDE_TABLEAM connection parameter,
which is honoured also in this commit.

Co-authored-by: Asim R P <apraveen@pivotal.io>