Properly transform aoco specific enconding clauses
This commit moves a lot of the aoco specific related functions from
parse_utilcmd to reloptions_gp. The intention is to try to separate as much as
possible greenplum specific behaviour from upstream and ultimately try to
utilize as much common code as possible.

Also in this commit, the parsing and analysis of encoding clauses happens later
in the processing. Specifically during DefineRelation. The reasoning for this
change is based to the fact that the accessMethod is not available during
parsing. This has caused the storage options to be processed once the
accessMethod became available, and the storage options are necessary for the
analysis of the enconding clauses. Since all these elements are tightly coupled
together, they are now moved closer together.

Co-authored-by: Asim R P <apraveen@pivotal.io>