Remove obsolete AddDefaultRelationAttributeOptions() function
Commit c29291bf68a in the iteration_REL_12 branch refactored the way that
enconding clauses are handled. It did not go as far as treating CTAS. This
commit completes the round.

Now the enconding options are handled by DefineRelation() in an attempt to
minimize the diff with upstream, bring the various greenplum specific options
closer together and reuse as much code as possible. With the refactor, the need
for AddDefaultRelationAttributeOptions() is now gone as the defaults are handled
and merged with specific options passed in a WITH clause during
transformAttributeEncoding() and added via AddRelationAttributeEncodings()

The commit addresses
```	duplicate key value violates unique
	constraint "pg_attribute_encoding_attrelid_attnum_index"```
errors.