Disable fsync_ao isolation2 test temporarily
The test is causing checkpointer to PANIC on mirror.  The code to
defer fsync of AO segment files to checkpointer on the mirror is
affected by the changes to SMGR layer.  We made a decision in
af05ebb26 to disable that code temporarily.  Let's also disable the
test until that is fixed.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>