Add tests for combinations of UPDATEing distribution/partition keys.
PostgreSQL now allows an UPDATE to move a row from one partition to
another. That works in GPDB now, as does UPDATEing a row from one segment
to another. But doing both in one UPDATE is currently broken. This adds
tests for all of these cases, with the correct expected output, to
demonstrate the broken case. The test is currently failing for the broken
case, and to make it pass, we need to fix the code.