Remove NOTICEs from expected output of ALTER TABLE SPLIT PARTITION
On master, ALTER TABLE SPLIT PARTITION prints NOTICEs like this:

    ALTER TABLE test_part SPLIT DEFAULT PARTITION
    START('2018-01-01'::date)  END( '2018-02-01'::date);
    NOTICE:  exchanged partition "p_overflow" of relation "test_part" with relation "pg_temp_32601"
    NOTICE:  dropped partition "p_overflow" for relation "test_part"
    NOTICE:  CREATE TABLE will create partition "test_part_1_prt_r28232684" for table "test_part"
    NOTICE:  CREATE TABLE will create partition "test_part_1_prt_p_overflow" for table "test_part"

On the merge branch, none of those NOTICEs are printed. I think that's good.
The first two notices are just noise. And misleading: the p_overflow
partition is only dropped as part of the internal dance of swapping it with
a temp table. It's still there after the command.

The "CREATE TABLE will create partition" NOTICEs are not printed by
CREATE TABLE anymore, either.

Remove the NOTICEs from expected output.