Revert plan change, changing test back the way it is on 'master'.
The access plan had changed from Hash Right Join to Hash Left Join earlier.
I'm not sure why. But now we're getting a Hash Right Join again. Not sure
why it changed back. But since it's now the same as on 'master', it's good
enough for me.

If this keeps flip-flopping in the future, maybe we need to change the test
to be more forgiving..