Remove test for an ancient subquery planning bug.
This test was for an old bug from 2009, fixed in GPDB 3.3:

    postgres=# create schema test;
    CREATE SCHEMA
    postgres=# select tablename from pg_tables where schemaname='test' and tablename not in( select partitiontablename from pg_partitions where partitionschemaname = 'test' );
    ERROR:  unrecognized node type: 231 (execAmi.c:217)

This test relied on pg_partitions table, which was removed with the merge
of upstream partitioning, but rather than try to write a replacement for
it, let's just remove it. Subquery processing has been refactored and
rewritten many times since, so it seems highly unlikely that this same bug
would re-appear. And without more details about the original bug (I couldn't
find the code change that fixed it), it's hard to tell what new test might
cover the same ground.