Remove test for an ancient subquery planning bug.
Original bug report (MPP-6327, duplicate of MPP-6419):

    ice_qa=# select * from ice_w.ice_lstg_desc_ins where icedt = (select partitionrangestart  FROM pg_partitions where  tablename='ice_lstg_desc_ins' and schemaname='ice_w' and partitionrank=1);

    ERROR:  Unexpected internal error: Segment process received signal SIGSEGV  (seg34 slice1 gauss005-3:50003 pid=25189)
    ERROR:  GPDB detected segment reconfiguration. (cdbgang.c:3478)

And simplified repro from the same:

    select * from test1 where '2009-12-12'::date = (select 'test'::text);

And here's the original fix for that:

    diff --git a/cdb-pg/src/backend/cdb/cdbmutate.c b/cdb-pg/src/backend/cdb/cdbmutate.c
    index 5b317fc1f3..2e50d7b75f 100644
    --- a/cdb-pg/src/backend/cdb/cdbmutate.c
    +++ b/cdb-pg/src/backend/cdb/cdbmutate.c
    @@ -2218,6 +2218,7 @@ static bool can_zap_result(Result *res)
            ListCell *lc;

            if(!res->plan.lefttree
    +      || res->plan.initPlan
              || !IsA(res->plan.lefttree, Append)
              || res->resconstantqual
              || res->hashFilter

The can_zap_result() function was removed in commit c613cabf34, so it seems
highly unlikely for this bug to reappear in the same form.

The test we had for that relied on the pg_partitions table, which was
removed with the merge, so we have to do something about it. Remove the
test.