Fix AO / AOCO same-xact TRUNCATE.
We failed to truncate segfile 0. Fix that.

Also, we no longer insert new rows to segfile 1, but segfile 0, in these
tests. Accept expected output.

In the passing, make the gpdiff "masking" less stringent in the test, so
that it doesn't mask out the segfile number in the filenames, only the
dboid and relfilenode.