Accept a few changes in plans, caused by partition pruning differences.
Postgres partition pruning knows that if there is a partition that matches
a qual, then there's no need to scan the default partition.

Also, the naming of the partitions changed slightly. Not sure if we're
going to try changing the naming on the merge branch, so that we'd get
the same names as before, but that's not the point of these tests.