Fix check for matching numsegments in partition hierarchy.
I retract my earlier FIXME comment about it being easy to support a
heterogenous hierarchy. It might be doable, but COPY routing to right
QE node would need to be changed, and probably planner changes too. So
while it might be possible, let's not try to cram that into the merge.