Modify dsp test sql and answer file
Modified test to not use subtransaction template as was not
necessary. Modifying test output based on some past modifications to
sql file.

The test has very small diff remaining now. Mainly two issues need to
be fixed:

1. Split partition should leverage access method guc to create new
tables (as per old expectation). I feel current method is better to
use current partitions access method.

2. Should not be allowed to set default_table_access_method GUC only
on master as value should be same for it on QD and QE. Not sure if
that expectation needs to meet or default_table_access_method can be
QD only guc.