Fix 'misc_sanity' failures caused by missing toast tables in catalogs.
Most of the extra GPDB catalog tables that it complains about are used
to store short system-generated strings, so we don't really need toast
tables for them. Perhaps we should create them anyway, to be tidy, but
let's not changes that on the merge branch.

I added a toast table for pg_extprotocol, however. It has a variable-
length ACLs field, which could be arbitrarily large if you do a lot of
GRANTs.