Accept change in output of test with repeated UPDATE WHERE CURRENT OF.
As far as I can see, the old behaviour memorized in the expected output
was incorrect. PostgreSQL has always returned the result we're now seeing
on the merge branch, which is different from the result you get on master.
The test had not been modified since 2012, when the test was added along
with support for WHERE CURRENT OF. I could not find any explanation in
git history or in test comments. I believe it was simply a bug when
WHERE CURRENT OF support was added, and no one noticed that the test
result was wrong.