Implement ALTER TABLE ... ALTER PARTITION ... SET DISTRIBUTED BY.
Doesn't seem very useful in practice, as you cannot change the
distribution policy of an individual partition. But it's pretty simple
to implement, so let's just do it. You can use it reorganize, so it's not
completely useless:

alter table mytab alter partition for (1) set with (reorganize=true)
  distributed by (i) ;

Also adjust test to not use RANK() addressing.