Convert END INCLUSIVE using the transformed Const value
Commit b23036066df7d4851855402c0dbed245c1e63f3b added initial support
for INCLUSIVE END for Integer and Date data types by constructing a "+1"
expression node and evaluate that. This patch instead directly operates
"+1" on the already transformed endConst->constvalue, which is more
straightforward and error pruning. Special handling for overflow are
also added if the INCLUSIVE END value is already the maximum value of
the data type.

For now, we only support the following 9 data types:

int2
int4
int8
date
time without time zone
time with time zone
interval
timestamp without time zone
timestamp with time zone

For the above types that support configurable precision, i.e, time (w/
and w/o time zone), interval, timestamp (w/ and w/o timezone), if a
precision is explicitly specified, END INCLUSIVE is also not allowed.

Additional tests are added for all of the supported types. Because this
is for legacy syntax, the intention is to support as less data types as
possible and only add support to additional data types if users require
in the future.

Now that we are not constructing the "+1" expression state for END
INCLUSIVE anymore as we do for EVERY, they no longer share the same
code, hence reverted the refactor code for that in commit
b23036066df7d4851855402c0dbed245c1e63f3b.