Remove obsolete window function planner test.
The original JIRA was:

    MPP-8946: Window node does not support Re-scan

    select schemaname, tablename, partitionschemaname,
    partitiontablename, partitionlevel from pg_partitions p1 where
    partitionlevel > 0 and partitionlevel < (select max(partitionlevel)
    from pg_partitions p2 where p2.tablename=p1.tablename);

    ERROR:  unrecognized node type: 232 (execAmi.c:222)

    Maybe execAmi.c:ExecReScan() needs a ExecShareInputScanReScan ?

    I may be able to come up with a workaround, but it could be hard for a
    customer..

This was fixed in GPDB 4.1.0.0. However, window function support was
completely rewritten in GPDB 5, replacing it with upstream code from
PostgreSQL 8.4. So it's not clear that this was doing anything interesting
anymore.

This relied on the window functions in the pg_partitions view, but that
view is gone now, and even if we were to re-implement it, it probably
would not happen to exercise the same codepath anymore. So let's just
remove the test.

(MPP-7073 was a duplicate of MPP-8946)