Fix PartitionSelector stuff.
This was surprisingly easy to make work with the upstream partitioning.
Planner required almost no changes. The main differences is that instead
of projecting a tuple using a target list, we evaluate a list of
expressions extract the partition keys from the input tuple. That works
better with the execution-time code to determine which partition the row
belongs to. We now use the upstream get_partition_for_tuple() function
for that.

This only fixes the "dynamic" way of choosing partitions, and only the
way the Postgres planner uses it. ORCA is still out of luck.

Still doesn't work with subpartitions. That's reflected by the remaining
'dpe' test failures.