Remove test query that used pg_partitions.
It seemed pretty unimportant for the test, so I just removed it.

Also remove a DROP TABLE that claimed that gpcheckcat had trouble with
the resulting table. Seems a bit dubious that it was ever possible to
end up in a situation like that, but surely with the full rewrite of
partitioning, it should no longer be an issue, whatever it was.