Use tableam api for relation size on AO/AOCO and general cleanup
This commit implements aoco_relation_size() and tries to use it where possible.
The tableam api interface does not give any option to switch over compressed or
uncompressed size. A study of the call sites revieled that the compressed site
is always desired, so that is what is implemented.

The logic is moved over from GetAOCSTotalBytes and is the equivelant for AO
relations, which copied GetAOTotalBytes. These two functions then became
obsolete and removed in favour of the tablam methods.

Along with it, RelationGuessNumberOfBlocks() has been removed in favour of
RelationGetNumberOfBlocks() from upstream. The upstream version recognised that
custom access methods might not use BLCKSZ wide fixed size block and implements
an approximation that was almost exact to Greenplum's.

Further cleanups on that front to follow.