Fix updatable cursors on partitioned tables.
This simplifies the code that deals with "simply updatable tables"
somewhat. Instead of digging the cursor's relation from the range table
in the executor, store the relation's OID in the planner, at the same time
that we determine that it is simply updatable in the first place. The
extractSimplyUpdatableRTEIndex() function was being confused by
subpartitions, because it expected there to be only one RTE with inh=true.

Per failures in 'portals_updatable' test.