Remove obsolete test.
Before the merge, partitioning was implemented as table inheritance,
and there was a CHECK constraint on each partition for the range (or list)
of values that belonged on that partition. This test tested that you could
not drop the automatically created CHECK constraint with ALTER TABLE DROP
CONSTRAINT. The upstream declarative partitioning doesn't rely on CHECK
constraints, the partition boundaries are stored separately, so there is
not automatic CHECK constraint that you could attempt to drop. And there is
a test in upstream 'alter_table' test that verifies that you cannot drop
a CHECK constraint on a partition that's inherited from the parent table.