Minor refactor of analyze targeting AO/AOCO relations
There had been a major refactor previously in analyze which included specific
functions for AO/AOCO (baa76536d). In iteration analyze had been remodelled and
the specific functions ignored and left dangling. This commit, properly removes
them.

Also the heap specific logic in analyze is bypassed, using the same mechanism as
before, though the tableam api is intended to be used as much as possible. One
ceveat, left as a GPDB_12_MERGE_FIXME, is that heap operates on blocks whereas
AO/AOCO operates on rows. However, there can be only uint32 blocks and there can
be uint64 rows. That means that rows greater than uint32 are not concidered in
the sample.

Finally, the last occurance of the AO/AOCO specific RelationGuessNumberOfBlocks
has been removed and substituted with RelationGuessNumberOfBlocksFromSize which
is also used by indexes.

Co-authored-by: Asim R P <pasim@vmware.com>