Analyze midlevel partitions like we used to.
At least I *think* it works the way it used to now.

Also, reverse the list of tables to analyze, when analyzing a partition
hierarchy, so that the leafs are analyzed before the parents. Otherwise
the GPDB code to merge leaf statistics won't be effective.

Fixes some failures in 'analyze' test.