Make ALTER TABLE .. SET TABLESPACE recursive
This has been the expected Greenplum behavior, which differs from
upstream Postgres.

In upstream Postgres, for a partitioned table foo, the command ALTER
TABLE foo SET TABLESPACE only sets the tablespace of table foo, the
existing partitions of table foo remain in the old tablespace, which
means this command for a partitioned table only rewrites the catalog and
does not rewrite the table data. If any new partitions are added after
performing the SET TABLESPACE command, the new partitions will reside in
the new tablespace.

In Greenplum, for a partitioned table foo, the command ALTER TABLE foo
SET TABLESPACE sets the tablespace of table foo and all the inheritor
partitions to the new tablespace recursively. For a multi-level
partitioned table, the legacy Greenplum command ALTER TABLE foo ALTER
partition foo_p1 SET TABLESPACE behaves the similar way, it sets the
tablespace of the middle level partitioned table foo_p1 and all of its
inheritors to the new tablespace.