Accept behavior difference with EXCHANGE PARTITION and CHECK constraints.
Previously, if a parttioned table had a CHECK constraint, and you
EXCHANGEd it with a table that didn't have the corresponding CHECK
constraint, the constraint was created and verified. Now, you get an
ERROR instead, and have to create the CHECK contraint manually. I think
that's OK, so accept that behavior, and adjust tests accordingly.