Use default permissions and Open/CloseTransientFile with 'savedtmredo.file'
The new initdb TAP tests complained if the file had non-default
permissions. I don't see any particular reason for the file to have
different permissions than all other files.

While we're at it, use Open/CloseTransientFile() for it. It doesn't matter
in practice: the file is only used at startup/shutdown, so leaks are not
a problem. And it seems unlikely that we would run out of file descriptors
either. But it seems like a good habit to use OpenTransientFile()
everywhere.