Create AO/AOCS aux relations earlier
Although it would be best if Aux relations were created during
heap_create_with_catalog() this is not as trivial due to snapshot visibility
issues. So the earliest place is in DefineRelation once the command counter has
been incremented.

That unifies a bit the locations of the call. Also a new function
NewRelationCreateAOAuxTables() is introduced as per the 94 fixme.

A small cleanup removes relevant missmerges in ProcessUtilitySlow().

Finally, this commit intentionally undoes commit 8bde459519951b, because it
would be preferable to catch these cases. If proven wrong, the commit will be
reinstated.

Kudos to Heikki for the discussions.