Use TTSOpsVirtual in AO/AOCO relations
Upstream TTSOpsVirtual do not implement a *_getsysattr method and error out in
this case. In order to not alter the behaviour of upstream, the VirtualOps where
overriden for AO/AOCO relations. Obviously this is not a suitable solution in
the long run. Since TTSOpsVirtual fit all other use cases quite naturally, it
makes sense to accept this deviation from upstream and allow for _getsysattr to
accept GpSegmentIdAttributeNumber input. In all other cases, the upstream
behaviour remains.