Minor refactor of aoco_scan_bitmap_next_tuple
In the recheck case, it was possible to fail to fetch a tuple and then try to
run the Qual expression on it. Given that AOCO is using Virtual Tuples, it means
that the expression would attempt to call the *get_someattrs method which would
error out.

This refactor, removes the unessary check and improves readability.

Co-authored-by: Heikki Linnakangas <hlinnakangas@pivotal.io>