Modify ALTER TABLE..ALTER PARTITION RENAME implementation
Old implementation recursed through the hierarchy via
RenameRelationInternal(). Due to which it was impossible to report
number of partition rename skips via one NOTICE for the command.

So, instead this commit modifies the logic to not recurse but instead
get all the OIDs using find_all_inheritors() and call
RenameRelationInternal() for each of them. This helps to easily count
how many partition renames are being skipped.

Important difference is only if prefix of partitions in heirarchy
matches with partition being renamed then we go ahead to rename
them. (Pre-GDPB7 logic if prefix had name of "immediate" parent then
only rename was performed.)

Co-authored-by: Alexandra Wang <walexandra@vmware.com>