Analyze test table, to get same plan as before.
Without the ANALYZE, the planner chooses a different plan, where the
Sort we're trying to test doesn't get executed.

In the passing, fix a crash I saw when running EXPLAIN ANALYZE on
this query. The instrumentation stats were missing for some reason. Not
sure how it happened - I guess it was because the Sort wasn't executed -
but it seems wise to be defensive in the show_sort_info() function, if the
stats are missing for any reason.