Change dml_on_root_locks_all_parts case to fit upstream code.
Before in tuple delete or update we have to upgrade sub-partition
table lock RowExclusiveLock to ExclusiveLock on initplan to prevent
deadlock. The reason is if another concurrency txn direct do some dml
in sub-partition table, it only apply a RowExclusiveLock for the
sub-partition table.

We do not need to worry it anymore. Now, the second transaction apply
ExclusiveLock for the sub table as upstream, so we can safely rid of our
gpdb private lock upgrade code.

change the test case to fit new stragety and rid of relavented dead
code.