Minor refactor to deal with extraction of columns for AOCO relations
The previous implementation in the current branch (iteration_REL_12) was
overriding struct ScanKeyData to pass down which columns to extract. It also
exposed a lot of implementation details that are better matched underneath the
tableam api. The tableam api is not mature enough yet to deal with this issue.
For this a couple of temporary extensions of it have been added.

These temporary extensions are intented to cover only these specific cases.
The only call sites, nodeSeqscan and nodeBitmapHeapscan are close enough but not
quite, in order to share the same method.

Since the AOCO specifics are now hidden, a couple of pubicly exposed functions
were able to be removed or replaced by local versions. Also some relevant minor
fixes are added.

Co-authored-by: Asim R P <pasim@vmware.com>