Accept potential as sync state in test output
Master/standby state comes out as "potential" in pg_stat_replication
because the GUC synchronous_standby_names is not defined.  Accept it
as expected answer as it means the master and standby are in sync.
Looks like the quorum commit 3901fd70 changed the behavior of
pg_stat_replication such that it started report "potential" if the
standby name is not included in the GUC.