Update index_constraint_naming_partition test
- Modified the sql validation queries in test to use
  pg_partition_tree() instead of pg_partitions

- Removed tests related to EXCHANGE PARTITIONS which had higher
  restrictions in pre-GPDB7

I didn't go through all the results checking the correctness of
dependencies reported and such. Trusting the upstream code to behave
correctly. I was on fence to delete this test but then finally decided
against it. If someone feels this test should be deleted I am okay
with it, for now I updated and decided not to loose the coverage.