Simplify aocs_beginscan() argument list
This function is now only called from the aoco am handler. In all cases the
snapshot used for accessing the metadata stored in aux tables, is the same as
the one used to access the relation data, unless that snapshot is Snapshot any
and then the transaction snapshot is used instead.

Push down that logic to aocs_beginscan and remove the chatter from the call
sites.