Disable triggers for statements
This referencing new table as ... for each statement feature lets the
trigger see a global view of what the statement did, not just one row at
a time.

Greenplum doesn't support it from the very beginning, but we never
disabled it, we just let the DDL succeed, and this kind of triggers work
or not by luck.

Since it's not supported, it most likely doesn't work, we have lots of
DDLs like this in the test not testing properly, disable it, and update
related test cases.