Cleanup bitmap scan for AOCO access method
Previously, in iteration_REL_12, the same scan descriptor would be used for all
scans. Which is largely correct, but misses an idiosyncracy of AO relations.
That is in order to retrieve a tuple based on the tid, a scan descriptor closer
resembling an index scan descriptor is needed. That very specific information
was exposed into the aocs scan descriptor and leaked to even deeper layers.

In this commit, a specific private scan descriptor is introduced. It is an
extension of a base table scan descriptor and for all the common upstream code
can be used interchangeably. It contains all the logic needed by the aocs
tableam handler and is not exposed anywhere else.

Also, in this commit, the implementation of _next_block and _next_tuple is
slightly optimized. The information that will be maintained between subsequent
_next_block calls, is now stored and used effectively in _next_tuple in order to
minimize branching. Additionally, _next_tuple has less work to do, as it is only
concerned with fetching the next tuple if found, besides the lazy initialization
of the fetch descriptor. That was not moved in the _next_block part, as it is
possible for the _next_tuple to be skipped.

Finally, the recheck logic, is removed from _next_tuple. It is a better fit
there, however the upstream code acknowledges that is violating the layers and
is responsible for the recheck case. AOCS is now following the upstream
paradigm.

Other minor cleanups include a clearer distinction of the recheck being used
also in the non lossy scans.

Co-authored-by: Asim R P <pasim@vmware.com>