Fix access method creation to use AO am handlers correctly
Previously, the access method OID was being used to determine the
behaviour of a relation.  For example, a custom access method will not
evaluate all options correctly during create table statement.

  create access method ao_row type table handler appendonly_tableam_handler;
  create table ao (a int, b int) using ao_row with (compresstype=zlib) distributed by (a);
  ERROR:  unrecognized parameter "compresstype"

This commit fixes it such that the access method handler OID is used
instead of the access method OID.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>