Fix a few unittests
All the unittests pertaining to AO and AOCO have been fixed.  A few
more tests are also fixed.  The compilation was found to fail due to
difficult to understand "symbol not found" errors for symbols defined.
We fixed these errors by using mocked version instead of the real
version of the applicable modules.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>