Abstract away AM specific initialization for AOCO during Alter Table
During alter table rewrite, the oldrel structure was abused to switch the tuple
descriptor in order to match what the AOCO api was expecting. Instead, change
the AOCO api to not need that change.

Now we lazily initialize AOCS scan datumstreams and the related info, based on
the slot provided during getnext tableam calls. That allowed for a minor clean
up of the API functions to remove unused arguments.

This allowed for a rewrite of the AOCScan structure to explicitly state the
internal column specific implementation details.

A minor cleanup of variables also followed, to consistently use the same types
and naming convensions everywhere.

While on it, a minor inconsistency and a bug where discovered in
open_next_scan_seg. It was possible for a scan with no projected attributes to
denote non empty segments and perform unnecessary segment and datumstream
operations. Also it was considered correct for a segment to contain tuples but
have zero length. This is now treated as an error.

Co-author-by: Asim R P <pasim@vmware.com>