Fix clash with table named 'parent' between tests.
I think the clash we saw in the pipeline was because both 'triggers' and
'inherit' tests used a table called 'parent', and they run in the same
parallel group.

Fix by moving the test from 'inherit'. It was a GPDB-added test, so it
didn't really belong in the upstream 'inherit' test, anyway.