Fix bogus plan with RowIdExpr.
The RowIdExpr needs to be evaluated on the outer side of the join,
not on the join node itself.

I thought we might have this bug on 'master', too, but I was not able
to reproduce it. I guess the logic to decide whether to use the "physical"
tlist is different enough on 'master'.

Also, the cost estimates are slightly different, so that you get a
different plan for this query than on 'master'. Expand the test so that
we test the new plan, but also the old plan.