Add fixme to note AOCO-specific code in ALTER TABLE ADD COLUMN implementation
It seems better to tackle the proposed refactor in the fixme on
master, after the iteration branch is merged.  A diff from upstream is
unavoidable in this case - either the table AM interface needs to
differ or the implemetation of ALTER TABLE ADD COLUMN needs to differ
(which is currently the case).  The change to table AM interface seems
significant enough to be made on master, instead of hiding it in a big
merge commit.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>