Properly fix the refactor in 3e09f4cb73
The commit 3e09f4cb73 tried to move the swap operation for
pg_appendonly entries to the appendoptimized table AM iplementation
for CLUSTER command.  However, that is not correct because swapping of
pg_appendonly entries is a generic operation, needed by commands such
as CLUSTER, REFRESH MATERIALIZED VIEW and ALTER TABLE commands that
rewrite the relation.  Therefore, reinstate the call to
SwapAppendonlyEntries to where the pg_class entries are swapped.

This change passess the tests that 3e09f4cb73 caused to fail.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>