Use table am handler to identify appendoptimized tables
Using table AM to identify if a relation is appendoptimized is not
sufficient.  New access methods can be defined that can still use
appendoptimized table AM handler functions.  Therefore, the table AM
handler function should be used to itendify appendoptimized tables.

This is the last commit in the series of commits intended to replace
the use of predefined AM OIDs with predfined handler OIDs.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>