Fix some unit test by inappropriate call will_assign_value
will_assign_value is a mock function to tell test framework if some
function invoke, assign a value to the implicit return variable.

If the return variable is a bool type, we should not give a rvalue to
will_assign_value, since the sizeof(bool) is 1 and when they mock
assign the value, the callback always thine the variable is a 4 bytes
variable.

Depends on compiler, some stack overflow may happen.

To fix this, replace all rvalue to a variable, then call will_assign_value
by variable.