Accept minor plan change.
The planner is still capable of producing the two-stage agg plan for
this query, it just considers it slightly more expensive now. You still
get it if you bump up gp_motion_cost_per_row. Not sure what changed the
estimate, but that's important for this test, so accept plan change.