Add a few fixmes related to attribute encodings
Added as fixmes because they reflect a state currently in master and the fixes
should be a bit more visible. The gist of the issue is that encodings are in
essence a per column extension of reloptions only for AOCO relations. The code
to set, alter and use them, should be common with reloptions yet it is not.
Also, there are discrepencies between different codepaths in the behaviour,
especially with note to precedence. Additionally, the code is scattered and can
be gathered under the same roof for readability, consistency and ease of
merging.

Co-authored-by: Asim R P <pasim@vmware.com>