Fix segfault in mock test.
The cmockery will_assign_macro() does "sizeof(value)" to determine how
much memory to copy through the given pointer. But in this case, at least
on my laptop sizeof(false) = 4, but sizeof(bool) is 1. That led to a
buffer over flow in the test. Fix by explicit cast to "bool".