Fix url stored in error log, when reading from gpfdist external table.
The URL stored in the 'filename' field of the error log was stored
incorrectly:

    postgres=# SELECT filename FROM gp_read_error_log('sreh_ext_err_tbl');
                                          filename
    ------------------------------------------------------------------------------------
     (null) [/home/heikki/git-sandbox-gpdb/master/src/test/regress/data/bad_data1.data]
     (null) [/home/heikki/git-sandbox-gpdb/master/src/test/regress/data/bad_data1.data]
    ...

On 5X_STABLE, the URL is stored in place of the '(null)'. This got broken
in 6X_STABLE. The root cause is that we no longer keep the 'url' in the
CopyState->filename field for gpfdist external tables.

Fixes https://github.com/greenplum-db/gpdb/issues/10542