Dust off DynamicSeqScan, episode one
This commit brings back minimal DynamicSeqScan support in the executor.
We can now execute simple plans from ORCA with a "Dynamic Table Scan"
node, although we cheat (for now): none of the pruning is done yet.

We also used this opportunity to delete a lot of old code.

Next up: add actual support for PartitionSelectors in the DynamicSeqScan
execution, while generating a "placeholder" partition selector (today we
hacked in a Result node).

Co-authored-by: Alexandra Wang <walexandra@vmware.com>