Make replace_line not suffer from overflow errors
During regress we generate files from a source, when we need to substitute
strings. Such an example is using different access methods but a single test
source file.

Long lines in source test files may suffer from overflows if the replacement
string is greater than the replaced string. This was the case in
selectinsert_while_vacuum.source. The way the code was for the substitutions,
the line buffer would overflow in the column case and would overwrite in the
line buffer for the row case which happened to be adjacent. As it happened the
overflow removed part of the statement and the remainder was syntactically
valid.

One possible solution would have been to reduce the line length in the offending
test case. However, the root cause would have remained. Instead, the
substitution code is now changed so that in case of overflows, the buffer
increases.

Co-authored-by: Asim R P <pasim@vmware.com>